+++ .phpunit.result.cache
@@ -1,1 +1,1 @@
-{"version":1,"defects":{"Warning":6,"CentralEntityTest::testAdd":4,"CentralEntityTest::testGetAvailableActions":3},"times":{"Warning":0.005,"CentralEntityTest::testAdd":0.005,"CentralEntityTest::testTransitionStatus":0.019,"CentralEntityTest::testGetAvailableActions":0}}
+{"version":1,"defects":{"Warning":6,"CentralEntityTest::testAdd":4,"CentralEntityTest::testGetAvailableActions":3,"CentralEntityTest::testTransitionStatus":3,"taskTest::testTransitionStatus":4,"taskTest::testGetAvailableActions":4},"times":{"Warning":0.005,"CentralEntityTest::testAdd":0.005,"CentralEntityTest::testTransitionStatus":0.035,"CentralEntityTest::testGetAvailableActions":0,"taskTest::testTransitionStatus":0.009,"taskTest::testGetAvailableActions":0}}
