+++ composer.json
@@ -4,7 +4,7 @@
     },
     "autoload": {
         "psr-4": {
-            "Class\\": "class"
+            "models\\": "models"
         }
     }
 }
